5.10:接入绑定设备和用户信息的界面，用户信息界面需要重写一下。有点乱

+++ New_iwear/Setting/InterfaceSelectionViewController/InterfaceSelectionModel/InterfaceSelectionModel.m
@@ -12,7 +12,7 @@
 
 - (NSString *)description
 {
-    return [NSString stringWithFormat:@"functionImageName == %@, functionName == %@, selectMode == %ld", _functionImageName, _functionName, _selectMode];
+    return [NSString stringWithFormat:@"functionImageName == %@, functionName == %@, selectMode == %ld", _functionImageName, _functionName, (unsigned long)_selectMode];
 }
 
 @end
